refactor: use src directory to declutter config files from src files
