--- frontend/app/(auth)/signin/page.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-export default function SignIn() {
-  return (
-    <div>signin</div>
-  );
-}
