--- frontend/app/page.js
@@ -1,6 +1,0 @@
-
-export default function Home() {
-  return (
-    <div>hi</div>
-  );
-}
