+++ frontend/src/app/(auth)/signin/page.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+export default function SignIn() {
+  return (
+    <div>signin</div>
+  );
+}
